refac: RecordImageService 예외 처리 리팩토링 및 추가 구현

+++ src/main/java/com/spot/spotserver/api/record/service/RecordImageService.java
@@ -2,8 +2,11 @@ package com.spot.spotserver.api.record.service;
 
 import com.spot.spotserver.api.record.domain.Record;
 import com.spot.spotserver.api.record.domain.RecordImage;
+import com.spot.spotserver.api.record.exception.RecordImageNotFoundException;
+import com.spot.spotserver.api.record.exception.RecordImageProcessingException;
 import com.spot.spotserver.api.record.repository.RecordImageRepository;
 import com.spot.spotserver.api.user.domain.User;
+import com.spot.spotserver.common.payload.ErrorCode;
 import com.spot.spotserver.common.s3.S3Service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -22,32 +25,47 @@ public class RecordImageService {
 
     @Transactional
     public String createRecordImage(MultipartFile recordImage, Record record, User user) throws IOException {
-        String imageUrl = this.s3Service.upload(recordImage, user.getNickname());
-        RecordImage newRecordImage = RecordImage.builder()
-                .url(imageUrl)
-                .record(record)
-                .build();
-        this.recordImageRepository.save(newRecordImage);
-        return imageUrl;
+        try {
+            String imageUrl = this.s3Service.upload(recordImage, user.getNickname());
+            RecordImage newRecordImage = RecordImage.builder()
+                    .url(imageUrl)
+                    .record(record)
+                    .build();
+            this.recordImageRepository.save(newRecordImage);
+            return imageUrl;
+        } catch (IOException e) {
+            throw new RecordImageProcessingException(ErrorCode.RECORD_IMAGE_PROCESSING_FAILED);
+        }
     }
 
     public String getRegionalThumbnailImage(Record record) {
-        RecordImage thumbnailImage = this.recordImageRepository.findFirstByRecordOrderByIdAsc(record).orElseThrow();
+        RecordImage thumbnailImage = this.recordImageRepository.findFirstByRecordOrderByIdAsc(record)
+                .orElseThrow(() -> new RecordImageNotFoundException(ErrorCode.RECORD_IMAGE_NOT_FOUND));
         return thumbnailImage.getUrl();
     }
 
     public List<String> getRecordImages(Record record) {
         List<RecordImage> recordImages = this.recordImageRepository.findAllByRecord(record);
+        if (recordImages.isEmpty()) {
+            throw new RecordImageNotFoundException(ErrorCode.RECORD_IMAGE_NOT_FOUND);
+        }
         return recordImages.stream().map((RecordImage::getUrl)).toList();
     }
 
     @Transactional
     public void deleteRecordImageById(Long recordImageId) {
+        if (!this.recordImageRepository.existsById(recordImageId)) {
+            throw new RecordImageNotFoundException(ErrorCode.RECORD_IMAGE_NOT_FOUND);
+        }
         this.recordImageRepository.deleteById(recordImageId);
     }
 
     @Transactional
     public void deleteRecordImage(Record record) {
+        List<RecordImage> recordImages = this.recordImageRepository.findAllByRecord(record);
+        if (recordImages.isEmpty()) {
+            throw new RecordImageNotFoundException(ErrorCode.RECORD_IMAGE_NOT_FOUND);
+        }
         this.recordImageRepository.deleteAllByRecord(record);
     }
 }
